# api-gateway/openapi.yaml
openapi: 3.0.0
info:
  title: TravelEase Api-docs
  description: "- Hotel Service: Built with Node.js and Express.js, utilizes MongoDB for its database.\n    \n- Auth Service: Utilizes Firebase for authentication and Firestore for the database.\n    \n- Booking Service: Another Node.js and Express.js service, interacting with a MongoDB database.\n    \n- Frontend: Developed using Angular."
  contact: {}
  version: '1.0'
servers:
- url: https://hotel-api-12oz.onrender.com/api/v1
  variables: {}
- url: https://booking-api-j0fx.onrender.com/api/v1
  variables: {}
paths:
...
  /rooms/book/ID:
    post:
      tags:
      - Room
      summary: book Room
      operationId: bookRoom
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/bookRoomRequest'
              - example:
                  checkInDate: 04-01-2024
                  checkOutDate: 04-14-2024
                  guests: 2
            example:
              checkInDate: 04-01-2024
              checkOutDate: 04-14-2024
              guests: 2
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: https://hotel-api-12oz.onrender.com/api/v1
        variables: {}
  /rooms/ID:
    get:
      tags:
      - Room
      summary: get room by Id
      operationId: getroombyId
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
      servers:
      - url: https://hotel-api-12oz.onrender.com/api/v1
        variables: {}
    patch:
      tags:
      - Room
      summary: update room
      operationId: updateroom
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/updateroomrequest'
              - example:
                  title: UPDATED room title
            example:
              title: UPDATED room title
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: https://hotel-api-12oz.onrender.com/api/v1
        variables: {}
    delete:
      tags:
      - Room
      summary: delete room
      operationId: deleteroom
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: https://hotel-api-12oz.onrender.com/api/v1
        variables: {}
...
  /create:
    post:
      tags:
      - booking
      summary: create booking
      operationId: createbooking
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/createbookingrequest'
              - example:
                  checkInDate: 09-18-2024
                  checkOutDate: 09-25-2024
                  guests: 2
                  user: testuser
                  roomId: 655ca5c959976c1c64f0872f
                  basePrice: 2000
                  ownerEarnedPrice: 1800
            example:
              checkInDate: 09-18-2024
              checkOutDate: 09-25-2024
              guests: 2
              user: testuser
              roomId: 655ca5c959976c1c64f0872f
              basePrice: 2000
              ownerEarnedPrice: 1800
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: https://booking-api-j0fx.onrender.com/api/v1
        variables: {}
components:
  schemas:
...
    bookRoomRequest:
      title: bookRoomRequest
      required:
      - checkInDate
      - checkOutDate
      - guests
      type: object
      properties:
        checkInDate:
          type: string
        checkOutDate:
          type: string
        guests:
          type: integer
          format: int32
      example:
        checkInDate: 04-01-2024
        checkOutDate: 04-14-2024
        guests: 2
    updateroomrequest:
      title: updateroomrequest
      required:
      - title
      type: object
      properties:
        title:
          type: string
      example:
        title: UPDATED room title
    createbookingrequest:
      title: createbookingrequest
      required:
      - checkInDate
      - checkOutDate
      - guests
      - user
      - roomId
      - basePrice
      - ownerEarnedPrice
      type: object
      properties:
        checkInDate:
          type: string
        checkOutDate:
          type: string
        guests:
          type: integer
          format: int32
        user:
          type: string
        roomId:
          type: string
        basePrice:
          type: integer
          format: int32
        ownerEarnedPrice:
          type: integer
          format: int32
      example:
        checkInDate: 09-18-2024
        checkOutDate: 09-25-2024
        guests: 2
        user: testuser
        roomId: 655ca5c959976c1c64f0872f
        basePrice: 2000
        ownerEarnedPrice: 1800
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...